# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Evig Direct Messages API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/templates": {
      "get": {
        "tags": [
          "templates"
        ],
        "summary": "List Templates",
        "operationId": "list_templates_api_v1_templates_get",
        "parameters": [
          {
            "name": "organization_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Organização atual. Padrão: organização única do ambiente.",
              "title": "Organization Id"
            },
            "description": "Organização atual. Padrão: organização única do ambiente."
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TemplateRead"
                  },
                  "title": "Response List Templates Api V1 Templates Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/templates/{template_id}": {
      "get": {
        "tags": [
          "templates"
        ],
        "summary": "Get Template",
        "operationId": "get_template_api_v1_templates__template_id__get",
        "parameters": [
          {
            "name": "template_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Template Id"
            }
          },
          {
            "name": "organization_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Organização atual. Padrão: organização única do ambiente.",
              "title": "Organization Id"
            },
            "description": "Organização atual. Padrão: organização única do ambiente."
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TemplateRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "templates"
        ],
        "summary": "Delete Template",
        "operationId": "delete_template_api_v1_templates__template_id__delete",
        "parameters": [
          {
            "name": "template_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Template Id"
            }
          },
          {
            "name": "organization_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Organização atual. Padrão: organização única do ambiente.",
              "title": "Organization Id"
            },
            "description": "Organização atual. Padrão: organização única do ambiente."
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/templates/drafts": {
      "post": {
        "tags": [
          "templates"
        ],
        "summary": "Create Draft",
        "operationId": "create_draft_api_v1_templates_drafts_post",
        "parameters": [
          {
            "name": "organization_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Organização atual. Padrão: organização única do ambiente.",
              "title": "Organization Id"
            },
            "description": "Organização atual. Padrão: organização única do ambiente."
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TemplateDraftCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TemplateRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/templates/{template_id}/draft": {
      "put": {
        "tags": [
          "templates"
        ],
        "summary": "Update Draft",
        "operationId": "update_draft_api_v1_templates__template_id__draft_put",
        "parameters": [
          {
            "name": "template_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Template Id"
            }
          },
          {
            "name": "organization_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Organização atual. Padrão: organização única do ambiente.",
              "title": "Organization Id"
            },
            "description": "Organização atual. Padrão: organização única do ambiente."
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TemplateDraftUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TemplateRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/templates/{template_id}/submit": {
      "post": {
        "tags": [
          "templates"
        ],
        "summary": "Submit Template",
        "operationId": "submit_template_api_v1_templates__template_id__submit_post",
        "parameters": [
          {
            "name": "template_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Template Id"
            }
          },
          {
            "name": "organization_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Organização atual. Padrão: organização única do ambiente.",
              "title": "Organization Id"
            },
            "description": "Organização atual. Padrão: organização única do ambiente."
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TemplateRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/templates/sync": {
      "post": {
        "tags": [
          "templates"
        ],
        "summary": "Synchronize Templates",
        "operationId": "synchronize_templates_api_v1_templates_sync_post",
        "parameters": [
          {
            "name": "organization_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Organização atual. Padrão: organização única do ambiente.",
              "title": "Organization Id"
            },
            "description": "Organização atual. Padrão: organização única do ambiente."
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SyncResult": {
        "properties": {
          "created": {
            "type": "integer",
            "title": "Created"
          },
          "updated": {
            "type": "integer",
            "title": "Updated"
          }
        },
        "type": "object",
        "required": [
          "created",
          "updated"
        ],
        "title": "SyncResult"
      },
      "TemplateCategory": {
        "type": "string",
        "enum": [
          "MARKETING",
          "UTILITY"
        ],
        "title": "TemplateCategory"
      },
      "TemplateDraftCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 512,
            "minLength": 3,
            "pattern": "^[a-z0-9_]+$",
            "title": "Name"
          },
          "display_name": {
            "type": "string",
            "maxLength": 512,
            "minLength": 3,
            "title": "Display Name"
          },
          "language": {
            "type": "string",
            "maxLength": 16,
            "title": "Language",
            "default": "pt_BR"
          },
          "category": {
            "$ref": "#/components/schemas/TemplateCategory"
          },
          "components": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Components"
          },
          "variable_schema": {
            "additionalProperties": true,
            "type": "object",
            "title": "Variable Schema"
          }
        },
        "type": "object",
        "required": [
          "name",
          "display_name",
          "category",
          "components"
        ],
        "title": "TemplateDraftCreate"
      },
      "TemplateDraftUpdate": {
        "properties": {
          "display_name": {
            "type": "string",
            "maxLength": 512,
            "minLength": 3,
            "title": "Display Name"
          },
          "category": {
            "$ref": "#/components/schemas/TemplateCategory"
          },
          "components": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Components"
          },
          "variable_schema": {
            "additionalProperties": true,
            "type": "object",
            "title": "Variable Schema"
          }
        },
        "type": "object",
        "required": [
          "display_name",
          "category",
          "components"
        ],
        "title": "TemplateDraftUpdate"
      },
...
      "TemplateRead": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "meta_template_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Meta Template Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "display_name": {
            "type": "string",
            "title": "Display Name"
          },
          "language": {
            "type": "string",
            "title": "Language"
          },
          "category": {
            "$ref": "#/components/schemas/TemplateCategory"
          },
          "status": {
            "$ref": "#/components/schemas/TemplateStatus"
          },
          "components": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Components"
          },
          "variable_schema": {
            "additionalProperties": true,
            "type": "object",
            "title": "Variable Schema"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "revision": {
            "type": "integer",
            "title": "Revision"
          },
          "rejection_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rejection Reason"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "meta_template_id",
          "name",
          "display_name",
          "language",
          "category",
          "status",
          "components",
          "variable_schema",
          "source",
          "revision",
          "rejection_reason",
          "created_at",
          "updated_at"
        ],
        "title": "TemplateRead"
      },
      "TemplateStatus": {
        "type": "string",
        "enum": [
          "DRAFT",
          "PENDING",
          "APPROVED",
          "REJECTED",
          "PAUSED",
          "DISABLED"
        ],
        "title": "TemplateStatus"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}